+++ images/demo/user-home/02 - Web3 Data (Ethereum trading example)/init-chapter-all.sh
@@ -4,15 +4,9 @@ set -e
 rm -rf .kamu
 kamu init
 
-# Pull from S3 for speed but then alias to IPFS
-kamu pull "https://s3.us-west-2.amazonaws.com/datasets.kamu.dev/net.rocketpool.reth.mint-burn" --as rocketpool.reth.mint-burn --no-alias
-kamu repo alias add --pull rocketpool.reth.mint-burn "ipns://net.rocketpool.reth.mint-burn.ipns.kamu.dev"
-
-kamu pull "https://s3.us-west-2.amazonaws.com/datasets.kamu.dev/com.cryptocompare.ohlcv.eth-usd" --as cryptocompare.ohlcv.eth-usd --no-alias
-kamu repo alias add --pull cryptocompare.ohlcv.eth-usd "ipns://com.cryptocompare.ohlcv.eth-usd.ipns.kamu.dev"
-
-kamu pull "https://s3.us-west-2.amazonaws.com/datasets.kamu.dev/co.alphavantage.tickers.daily.spy" --as alphavantage.tickers.daily.spy --no-alias
-kamu repo alias add --pull alphavantage.tickers.daily.spy "ipns://co.alphavantage.tickers.daily.spy.ipns.kamu.dev"
+kamu pull "https://s3.us-west-2.amazonaws.com/datasets.kamu.dev/net.rocketpool.reth.mint-burn"
+kamu pull "https://s3.us-west-2.amazonaws.com/datasets.kamu.dev/com.cryptocompare.ohlcv.eth-usd"
+kamu pull "https://s3.us-west-2.amazonaws.com/datasets.kamu.dev/co.alphavantage.tickers.daily.spy"
 
 kamu add -r datasets/